+++ tests/test_voxel_graph.h
@@ -27,6 +27,7 @@ void test_voxel_graph_hash();
 #endif
 void test_voxel_graph_issue471();
 void test_voxel_graph_unused_single_texture_output();
+void test_voxel_graph_spots2d_optimized_execution_map();
 
 } // namespace zylann::voxel::tests